Remove legacy COE pages and refactor
